load cms data before loading screen
